+++ radiotaphdr.h
@@ -5,8 +5,8 @@
 struct RadiotapHdr {
 	uint8_t ver_;
 	uint8_t pad_;
-	uint8_t len_;
-	uint8_t present_;
+	uint16_t len_;
+	uint32_t present_;
 };
 typedef RadiotapHdr *PRadiotapHdr;
 #pragma pack(pop)
